+++ internal/pkg/instrumentation/bpf/net/http/client/bpf_bpfel_arm64.go
@@ -13,13 +13,14 @@ import (
 )
 
 type bpfHttpRequestT struct {
-	StartTime uint64
-	EndTime   uint64
-	Sc        bpfSpanContext
-	Psc       bpfSpanContext
-	Method    [10]int8
-	Path      [100]int8
-	_         [2]byte
+	StartTime  uint64
+	EndTime    uint64
+	Sc         bpfSpanContext
+	Psc        bpfSpanContext
+	StatusCode uint64
+	Method     [10]int8
+	Path       [100]int8
+	_          [2]byte
 }
 
 type bpfSliceArrayBuff struct{ Buff [1024]uint8 }
@@ -70,21 +71,22 @@ type bpfSpecs struct {
 //
 // It can be passed ebpf.CollectionSpec.Assign.
 type bpfProgramSpecs struct {
-	UprobeHttpClientDo        *ebpf.ProgramSpec `ebpf:"uprobe_HttpClient_Do"`
-	UprobeHttpClientDoReturns *ebpf.ProgramSpec `ebpf:"uprobe_HttpClient_Do_Returns"`
+	UprobeTransportRoundTrip        *ebpf.ProgramSpec `ebpf:"uprobe_Transport_roundTrip"`
+	UprobeTransportRoundTripReturns *ebpf.ProgramSpec `ebpf:"uprobe_Transport_roundTrip_Returns"`
 }
 
 // bpfMapSpecs contains maps before they are loaded into the kernel.
 //
 // It can be passed ebpf.CollectionSpec.Assign.
 type bpfMapSpecs struct {
-	AllocMap                  *ebpf.MapSpec `ebpf:"alloc_map"`
-	Events                    *ebpf.MapSpec `ebpf:"events"`
-	GolangMapbucketStorageMap *ebpf.MapSpec `ebpf:"golang_mapbucket_storage_map"`
-	HttpEvents                *ebpf.MapSpec `ebpf:"http_events"`
-	SliceArrayBuffMap         *ebpf.MapSpec `ebpf:"slice_array_buff_map"`
-	TrackedSpans              *ebpf.MapSpec `ebpf:"tracked_spans"`
-	TrackedSpansBySc          *ebpf.MapSpec `ebpf:"tracked_spans_by_sc"`
+	AllocMap                   *ebpf.MapSpec `ebpf:"alloc_map"`
+	Events                     *ebpf.MapSpec `ebpf:"events"`
+	GolangMapbucketStorageMap  *ebpf.MapSpec `ebpf:"golang_mapbucket_storage_map"`
+	HttpClientUprobeStorageMap *ebpf.MapSpec `ebpf:"http_client_uprobe_storage_map"`
+	HttpEvents                 *ebpf.MapSpec `ebpf:"http_events"`
+	SliceArrayBuffMap          *ebpf.MapSpec `ebpf:"slice_array_buff_map"`
+	TrackedSpans               *ebpf.MapSpec `ebpf:"tracked_spans"`
+	TrackedSpansBySc           *ebpf.MapSpec `ebpf:"tracked_spans_by_sc"`
 }
 
 // bpfObjects contains all objects after they have been loaded into the kernel.
@@ -106,13 +108,14 @@ func (o *bpfObjects) Close() error {
 //
 // It can be passed to loadBpfObjects or ebpf.CollectionSpec.LoadAndAssign.
 type bpfMaps struct {
-	AllocMap                  *ebpf.Map `ebpf:"alloc_map"`
-	Events                    *ebpf.Map `ebpf:"events"`
-	GolangMapbucketStorageMap *ebpf.Map `ebpf:"golang_mapbucket_storage_map"`
-	HttpEvents                *ebpf.Map `ebpf:"http_events"`
-	SliceArrayBuffMap         *ebpf.Map `ebpf:"slice_array_buff_map"`
-	TrackedSpans              *ebpf.Map `ebpf:"tracked_spans"`
-	TrackedSpansBySc          *ebpf.Map `ebpf:"tracked_spans_by_sc"`
+	AllocMap                   *ebpf.Map `ebpf:"alloc_map"`
+	Events                     *ebpf.Map `ebpf:"events"`
+	GolangMapbucketStorageMap  *ebpf.Map `ebpf:"golang_mapbucket_storage_map"`
+	HttpClientUprobeStorageMap *ebpf.Map `ebpf:"http_client_uprobe_storage_map"`
+	HttpEvents                 *ebpf.Map `ebpf:"http_events"`
+	SliceArrayBuffMap          *ebpf.Map `ebpf:"slice_array_buff_map"`
+	TrackedSpans               *ebpf.Map `ebpf:"tracked_spans"`
+	TrackedSpansBySc           *ebpf.Map `ebpf:"tracked_spans_by_sc"`
 }
 
 func (m *bpfMaps) Close() error {
@@ -120,6 +123,7 @@ func (m *bpfMaps) Close() error {
 		m.AllocMap,
 		m.Events,
 		m.GolangMapbucketStorageMap,
+		m.HttpClientUprobeStorageMap,
 		m.HttpEvents,
 		m.SliceArrayBuffMap,
 		m.TrackedSpans,
@@ -131,14 +135,14 @@ func (m *bpfMaps) Close() error {
 //
 // It can be passed to loadBpfObjects or ebpf.CollectionSpec.LoadAndAssign.
 type bpfPrograms struct {
-	UprobeHttpClientDo        *ebpf.Program `ebpf:"uprobe_HttpClient_Do"`
-	UprobeHttpClientDoReturns *ebpf.Program `ebpf:"uprobe_HttpClient_Do_Returns"`
+	UprobeTransportRoundTrip        *ebpf.Program `ebpf:"uprobe_Transport_roundTrip"`
+	UprobeTransportRoundTripReturns *ebpf.Program `ebpf:"uprobe_Transport_roundTrip_Returns"`
 }
 
 func (p *bpfPrograms) Close() error {
 	return _BpfClose(
-		p.UprobeHttpClientDo,
-		p.UprobeHttpClientDoReturns,
+		p.UprobeTransportRoundTrip,
+		p.UprobeTransportRoundTripReturns,
 	)
 }